HSEARCH-2581 Give more specific names to Elasticsearch transport threads
This should make them more easily identifiable. Also, this proves that
we actually can customize these names, which will be required in the
future (when we'll have multiple clients in particular).
